06_fine_tune_llm's/requirements.txt
3datasets==3.6.0
peft==0.16.0
transformers==4.53.1
